+++ src/FluxFlow.Components.State/StateComponentTypes.cs
@@ -1,0 +1,8 @@
+using FluxFlow.Engine.Definitions;
+
+namespace FluxFlow.Components.State;
+
+public static class StateComponentTypes
+{
+    public static readonly NodeType Reducer = new("state.reducer");
+}